[Update] Refactor code Moved signal for starting the camera thread in the camera app signals
