+++ mask_detect/accounts/signals.py
@@ -15,9 +15,5 @@ def add_employee_to_company(sender, instance, **kwargs):
         instance.company = company.first()
 
 
-@receiver(user_logged_in)
-def post_login(sender, user, request, **kwargs):
-    camera_thread = CameraThread(user)
-    camera_thread.setDaemon(True)
-    camera_thread.start()
+
