+++ mask_detect/camera/apps.py
@@ -4,3 +4,6 @@ from threading import Thread
 
 class CameraConfig(AppConfig):
     name = 'camera'
+
+    def ready(self):
+        import camera.signals
